# Conflicts:
#	.gitignore
#	README.md
#	app.js
#	config.js
#	graphql/schema.graphql
#	idps/google.js
#	idps/index.js
#	package-lock.json
#	package.json
#	routes/auth.js
#	test/init-test-data.js

+++ test/init-test-data.js
@@ -1,40 +1,27 @@
-require('dotenv').config()
+const config = require('../config');
 const fs = require('fs')
 const yaml = require('js-yaml')
-const {executeQuery} = require("../database/neo4j-service");
-const {registerUser} = require("../database/data-interface");
-const fileData = fs.readFileSync('./yaml/test-data.yaml', 'utf8')
+const {wipeDatabase} = require("../data-management/neo4j-service");
+const {registerUser} = require("../data-management/data-interface");
+
+//Read user info data
+const fileData = fs.readFileSync(config.DATA_FILE, 'utf8')
 const data = yaml.load(fileData)
 const users = data['users']
 
-wipeDatabase()
-loadTestData(users)
-process.exit(0)
+loadTestData(users).then(r => process.exit(0))
 
-function wipeDatabase() {
-    if (process.env.TEST_LOADING_MODE === "overwrite"){
+async function loadTestData(users) {
+    if (config.DATA_LOADING_MODE === "overwrite") {
         console.log("Wiping database");
-        executeQuery({}, `MATCH (n) OPTIONAL MATCH (n)-[r]-() DELETE r,n`, {})
-        return
+        await wipeDatabase()
     }
-    else{
-        return
-    }
-}
-
-function loadTestData(users) {
-    for (const user of users){
-        let numLoaded = 0;
+    let numLoaded = 0;
+    for (const user of users) {
         const parameters = {}
         parameters['userInfo'] = user
-        try{
-            registerUser(parameters)
-            numLoaded++
-        }
-        catch (error) {
-            console.error(error);
-        }
-        console.log("Loaded "+numLoaded+" test users");
+        console.log(await registerUser(parameters))
+        numLoaded++
     }
-    return
+    console.log(numLoaded + " users loaded")
 }
